Rename to BaselineFacade
