+++ detekt-cli/src/test/kotlin/io/gitlab/arturbosch/detekt/cli/BaselineFacadeTest.kt
@@ -1,21 +1,21 @@
 package io.gitlab.arturbosch.detekt.cli
 
 import io.gitlab.arturbosch.detekt.cli.baseline.BaselineFormat
-import io.gitlab.arturbosch.detekt.cli.out.DetektBaselineFormat
+import io.gitlab.arturbosch.detekt.cli.baseline.BaselineFacade
 import org.junit.jupiter.api.Test
 import java.nio.file.Files
 
 /**
  * @author Artur Bosch
  */
-internal class DetektBaselineFormatTest {
+internal class BaselineFacadeTest {
 
 	private val path = Files.createTempDirectory("baseline_format")
 	private val fullPath = path.resolve("baseline.xml")
 
 	@Test
 	fun create() {
-		val format = DetektBaselineFormat(fullPath)
+		val format = BaselineFacade(fullPath)
 		format.create(emptyList())
 		BaselineFormat.read(fullPath)
 	}
